pyproject.toml
[project]
name = "educationa-ban"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agents>=1.4.0",
    "fastapi>=0.118.0",
    "openai>=2.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
